requirements.txt
absl-py==0.11.0
altgraph==0.10.2
appdirs==1.4.4
asn1crypto==0.24.0
astor==0.8.1
attrs==20.3.0
Automat==20.2.0
backports-abc==0.5
backports.functools-lru-cache==1.6.1
backports.weakref==1.0.post1
bdist-mpkg==0.5.0
beautifulsoup4==4.9.3
bleach==1.5.0
bonjour-py==0.3
cachetools==3.1.1
certifi==2020.12.5
cffi==1.12.2
chardet==4.0.0
chromedriver==2.24.1
click==7.1.2
colorama==0.4.4
configparser==4.0.2
constantly==15.1.0
contextlib2==0.6.0.post1
crayons==0.4.0
cryptography==2.6.1
cssselect==1.1.0
distlib==0.3.1
enum34==1.1.6
filelock==3.0.12
flake8==3.8.4
Flask==1.1.2
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.17.1
futures==3.3.0
gast==0.2.2
google-auth==1.27.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.36.1
h5py==2.10.0
html5lib==0.9999999
hyperlink==21.0.0
idna==2.10
importlib-metadata==2.0.0
importlib-resources==3.3.1
incremental==17.5.0
ipaddress==1.0.22
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
lxml==4.6.2
macholib==1.5.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==1.3.1
mccabe==0.6.1
mock==3.0.5
modulegraph==0.10.4
mongo==0.2.0
nltk==3.5
nose==1.3.7
numpy==1.16.6
oauthlib==3.1.0
opt-einsum==2.3.2
pandas==0.24.2
parsel==1.6.0
pathlib2==2.3.5
peewee==3.14.0
Pillow==6.2.2
Protego==0.1.16
protobuf==3.15.5
psycopg2-binary==2.8.6
py2app==0.7.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.19
PyDispatcher==2.0.5
pyee==7.0.4
pyflakes==2.2.0
PyHamcrest==1.10.1
pymongo==3.11.3
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==19.0.0
pyparsing==2.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2013.7
queuelib==1.5.0
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.5
scandir==1.10.0
scipy==1.2.2
Scrapy==1.8.0
selenium==3.141.0
service-identity==18.1.0
singledispatch==3.4.0.3
six==1.12.0
slack==0.0.2
slackapi==0.0.3
slackclient==1.3.2
slackeventsapi==2.2.1
slacksdk==0.0.1a0
soupsieve==1.9.6
tensorboard==2.1.0
tensorflow==1.5.0
tensorflow-estimator==2.1.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
tornado==5.1.1
tqdm==4.59.0
Twisted==20.3.0
typing==3.7.4.3
urllib3==1.26.2
virtualenv==20.2.2
w3lib==1.22.0
webdriver-manager==2.4.0
websocket-client==0.54.0
Werkzeug==1.0.1
wrapt==1.12.1
xattr==0.6.4
zipp==1.2.0
zope.interface==5.2.0
